backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
stripe==10.0.0
openai==1.50.0
weasyprint==62.3
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.2
PyJWT==2.9.0
bcrypt==4.2.0
